oppgavepakke/requirements.txt
Flask
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
opentelemetry-exporter-prometheus
prometheus-client
werkzeug
